# Conflicts:
#	CHANGELOG.md

+++ stubs/tests/inertia/DeleteAccountTest.php
@@ -4,6 +4,7 @@ namespace Tests\Feature;
 
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Laravel\Jetstream\Features;
 use Tests\TestCase;
 
 class DeleteAccountTest extends TestCase
@@ -12,6 +13,10 @@ class DeleteAccountTest extends TestCase
 
     public function test_user_accounts_can_be_deleted()
     {
+        if (! Features::hasAccountDeletionFeatures()) {
+            return $this->markTestSkipped('Account deletion is not enabled.');
+        }
+
         $this->actingAs($user = User::factory()->create());
 
         $response = $this->delete('/user', [
@@ -23,6 +28,10 @@ class DeleteAccountTest extends TestCase
 
     public function test_correct_password_must_be_provided_before_account_can_be_deleted()
     {
+        if (! Features::hasAccountDeletionFeatures()) {
+            return $this->markTestSkipped('Account deletion is not enabled.');
+        }
+
         $this->actingAs($user = User::factory()->create());
 
         $response = $this->delete('/user', [
